requirements.txt
python-binance==1.0.17
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
ta==0.10.2
dash==2.14.0
plotly==5.18.0
praw==7.7.1
tradingview-ta==3.3.0
numpy>=1.24.3
scikit-learn>=1.3.0